+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Compatibility:
 * Support `Proc#initialize_{dup,copy}` for subclasses (#3681, @rwstauner).
 * Remove deprecated `Encoding#replicate` method (#3681, @rwstauner).
 * Add `ObjectSpace::WeakMap#delete` (#3681, @andrykonchin).
+* `Kernel#lambda` with now raises `ArgumentError` when given a non-lambda, non-literal block (#3681, @Th3-M4jor).
 
 Performance: